+++ src/core/message/middlewares/format.js
@@ -1,0 +1,5 @@
+/**
+ * formats a message.
+ * @param {Message} msg 
+ */
+export default function format(msg) { msg.param.message = unescape(msg.param.message.replace(/&#(\d+);/g, (match, str) => '%' + parseInt(str).toString(16))) }