+++ src/modules/osubot/commands/_messages.js
@@ -1,5 +1,5 @@
 export default {
-    QUERY_BIND_FAIL: '你还没有绑定你的osu!id。\n使用 `-bind <id>\' 来绑定（*一定*要去掉两边的尖括号<>），\n如果用户名有空格请将用户名*整个*用英文引号 "" 括起来！',
+    QUERY_BIND_FAIL: '你还没有绑定你的osu!id。\n使用 `-bind <id>\' 来绑定（*一定*要去掉两边的尖括号<>）。',
     QUERY_NET_FAIL: '用户不存在/最近没玩过！',
     QUERY_CANVAS_FAIL: '请求太频繁，请稍后重试！',
     BP_ARGS_FAIL: '请指定一个bp序号(1-100)！',